+++ src/New/Type/ShieldTypeClass.cpp
@@ -42,6 +42,9 @@ void ShieldTypeClass::LoadFromINI(CCINIClass* pINI)
 
 	this->BreakAnim.Read(exINI, pSection, "BreakAnim");
 	this->HitAnim.Read(exINI, pSection, "HitAnim");
+
+	this->AbsorbPercent.Read(exINI, pSection, "AbsorbPercent");
+	this->PassPercent.Read(exINI, pSection, "PassPercent");
 }
 
 template <typename T>
@@ -62,6 +65,8 @@ void ShieldTypeClass::Serialize(T& Stm)
 		.Process(this->IdleAnim)
 		.Process(this->BreakAnim)
 		.Process(this->HitAnim)
+		.Process(this->AbsorbPercent)
+		.Process(this->PassPercent)
 		;
 }